+++ r_gl.h
@@ -48,7 +48,7 @@ public:
 	virtual void Begin2D() override;
 	virtual void End2D() override;
 	virtual void SetFont(CFont* font) override;
-	virtual void DrawText(const std::string& text, int x, int y) override;
+	virtual void DrawText(const std::string& text, float x, float y) override;
 	virtual void RenderEnd(void)						override;
 	virtual void SetWindowTitle(char* windowTitle) override;
